refactor: split up models.py and qtype.py
